+++ lib/actions/create_payment.d.ts
@@ -2,4 +2,5 @@ import { EVENT, Config } from "../types";
 export declare const create_payment: (event: EVENT, options: Config) => Promise<{
     created: number;
     failed: number;
+    failed_msg: never[];
 }>;