+++ lib/index.d.ts
@@ -1,6 +1,7 @@
 export declare const Actions: (event: import("./types.js").EVENT, options: import("./types.js").Config) => Promise<{
     created: number;
     failed: number;
+    failed_msg: never[];
 }>;
 export declare const ActionsList: import("./types.js").Action[];
 export declare const Commands: (CommandEvent: import("./types.js").CommandEvent) => Promise<import("./types.js").Result | {